requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
ruff==0.0.287
autopep8==2.0.4